Removed any mentions of Hecate and fixed spect*r*ogram typos
