+++ models.py
@@ -7,18 +7,11 @@ from typing import List, Optional, TypedDict
 # These are the types of output this worker (possibly) provides (depending on configuration)
 class OutputType(Enum):
     KEYFRAMES = "keyframes"  # produced by keyframe_extraction.py
-    METADATA = "metadata"  # produced by hecate.py
+    METADATA = "metadata" # produced by scenedetect.py
     PROVENANCE = "provenance"  # produced by provenance.py
     SPECTOGRAMS = "spectograms"  # produced by spectogram.py
     AUDIO = "audio"  # produced by spectogram.py
     SPECTOGRAM_IMAGES = "spectogram_images"  # produced by spectogram.py
-
-
-# Hecate outputs these files into OutputType.METADATA
-class HecateOutput(Enum):
-    KEYFRAME_INDICES = "keyframes_indices.txt"
-    KEYFRAMES_TIMESTAMPS = "keyframes_timestamps_ms.txt"
-    SHOT_BOUNDARIES = "shot_boundaries_timestamps_ms.txt"
 
 
 class ScenedetectOutput(Enum):
