+++ tests/data/spectrograms/test_example_output/0.npz
