+++ tests/data/spectrograms/test_example_output/1.npz
