+++ tests/data/spectrograms/test_example_output/2.npz
